Add a more explicit note on SSR
There is already a `How to implement server-side rendering` link to a very nice article. However, we should still be explicit regarding not to use `ConnectedRouter` on server-side.

Indeed, in one of our applications, we happen to be running with  the `ConnectedRouter` on server-side and it was running fine. However, after updating `react-redux` to `7.1` (from `7.0`), the app was starting to break. This was due to the fact that we were using the `context` prop like the `staticContext` prop of `StaticRouter` 😬 Of course, all our logic using the `StaticContext` was also silently breaking ^^